Add setup base classes and events to Pages
