+++ MvvmCross-Forms/MvvmCross.Forms.iOS/MvvmCross.Forms.iOS.csproj
@@ -41,6 +41,7 @@
     <Compile Include="MvxFormsApplicationDelegate.cs" />
     <Compile Include="MvxPageRenderer.cs" />
     <Compile Include="Views\Renderers\MvxImageViewRenderer.cs" />
+    <Compile Include="MvxFormsIosSetup.cs" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="System" />
@@ -102,6 +103,10 @@
     <ProjectReference Include="..\..\MvvmCross\Platform\iOS\MvvmCross.Platform.iOS.csproj">
       <Project>{BBBCD850-E7CB-4B6C-86D9-CE01F2B1C81C}</Project>
       <Name>MvvmCross.Platform.iOS</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\..\MvvmCross\Core\Localization\MvvmCross.Localization.csproj">
+      <Project>{D89351C1-D48F-4AD0-A0B9-353A93425AB7}</Project>
+      <Name>MvvmCross.Localization</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="..\..\packages\Xamarin.Forms.2.3.4.231\build\portable-win+net45+wp80+win81+wpa81+MonoAndroid10+Xamarin.iOS10+xamarinmac20\Xamarin.Forms.targets" Condition="Exists('..\..\packages\Xamarin.Forms.2.3.4.231\build\portable-win+net45+wp80+win81+wpa81+MonoAndroid10+Xamarin.iOS10+xamarinmac20\Xamarin.Forms.targets')" />
