+++ TestProjects/Forms/MvxBindings/MvxBindingsExample.UWP/Setup.cs
@@ -15,13 +15,10 @@ using ExampleApp = MvxBindingsExample.App;
 
 namespace MvxBindingsExample.UWP
 {
-    public class Setup : MvxWindowsSetup
+    public class Setup : MvxFormsWindowsSetup
     {
-        private readonly LaunchActivatedEventArgs _launchActivatedEventArgs;
-
-        public Setup(XamlControls.Frame rootFrame, LaunchActivatedEventArgs e) : base(rootFrame)
+        public Setup(XamlControls.Frame rootFrame, LaunchActivatedEventArgs e) : base(rootFrame, e)
         {
-            _launchActivatedEventArgs = e;
         }
 
         protected override IMvxApplication CreateApp()
@@ -33,36 +30,5 @@ namespace MvxBindingsExample.UWP
         {
             return new DebugTrace();
         }
-
-        protected override IMvxWindowsViewPresenter CreateViewPresenter(IMvxWindowsFrame rootFrame)
-        {
-            Forms.Init(_launchActivatedEventArgs);
-
-            var xamarinFormsApp = new MvxFormsApp();
-            var presenter = new MvxFormsUwpPagePresenter(  rootFrame, xamarinFormsApp);
-            Mvx.RegisterSingleton<IMvxViewPresenter>(presenter);
-
-            return presenter;
-        }
-
-        protected override void InitializeLastChance()
-        {
-            InitializeBindingBuilder();
-
-            base.InitializeLastChance();
-        }
-
-        protected virtual void InitializeBindingBuilder()
-        {
-            MvxBindingBuilder bindingBuilder = CreateBindingBuilder();
-
-            bindingBuilder.DoRegistration();
-        }
-
-        protected virtual MvxBindingBuilder CreateBindingBuilder()
-        {
-            return new MvxFormsBindingBuilder();
-        }
-
     }
 }
