+++ TestProjects/Forms/MvxBindings/MvxBindingsExample/Pages/MainPage.xaml.cs
@@ -1,6 +1,9 @@
-﻿namespace MvxBindingsExample.Pages
+﻿using MvvmCross.Forms.Core;
+using MvxBindingsExample.ViewModels;
+
+namespace MvxBindingsExample.Pages
 {
-    public partial class MainPage
+    public partial class MainPage : MvxContentPage<MainViewModel>
     {
         public MainPage()
         {
